refactor: remove baseService shim
