+++ src/logic/operationInterpreter.js
@@ -4,7 +4,7 @@
 // -----------------------------------------------------------------------------
 
 import { resolvePlaceholders } from '../utils/contextUtils.js';
-import { BaseService } from '../utils/baseService.js';
+import { BaseService } from '../utils/serviceBase.js';
 import { getNormalizedOperationType } from '../utils/operationTypeUtils.js';
 
 /** @typedef {import('../../data/schemas/operation.schema.json').Operation} Operation */
