--- src/utils/baseService.js
@@ -1,1 +1,0 @@
-export { BaseService } from './serviceBase.js';
